Find the data library in the script directory
Prior to this commit, the "data" directory is looked for in the working
directory. As a consequence, cgHeliParm has to be called from its own
directory, or the data directory has to be copied alongside the
trajectory.

This commit introduce the SCRIPT_DIRECTORY variable at the beginning of
all the cgHeliParm scripts. This variable stores the path to the
directory were chHeliParm sits. Thanks to this variable, the data
directory can be looked for next to the script rather than in the
working directory; so cgHeliParm can be called from any directory.

This assumes that the content of data is not to change on a
trajectory-per-trajectory basis and that it is inherently attached to
the program rather than to the trajectory.